Hexadecimal is never an empty string (#593)
Similarly to UUIDs, a hexadecimal can never be empty, this PR documents
it as such.

More and more libraries are starting to enforce stricter typing, I was
dealing with a library that required a `non-empty-string` which
hexadecimal isn't (is, but isn't documented that it is) which requires
either baselining false positives are adding redundant assertions.